Split out getDomain
